Preprocess JSON in LeadProviderApiSpecification
Previously I tried to preprocess the JSON in the middleman build, but it turns
out we serve the JSON files from the application rather than the API
docs/middleman project.

Hook into `Preprocessor` as part of the `LeadProviderApiSpecification`
library.
