+++ app/lib/lead_provider_api_specification.rb
@@ -15,6 +15,10 @@ class LeadProviderApiSpecification
   end
 
   def self.spec(version = CURRENT_VERSION)
-    YAML.load_file("swagger/#{version}/api_spec.json")
+    swagger_file = "swagger/#{version}/api_spec.json"
+
+    Preprocessor.new(swagger_file).preprocess!
+
+    YAML.load_file(swagger_file)
   end
 end
